ADD storybookOptions to the input for chromatic
