+++ bin/storybook/extract.js
@@ -1,30 +1,3 @@
-const CHROMATIC_PARAMETERS = [
-  'viewports',
-  'delay',
-  'disable',
-  'noScroll',
-  'diffThreshold',
-  'pauseAnimationAtEnd',
-];
-
-function specFromStory({ id, kind, name, parameters: { chromatic } = {} }) {
-  const param = value => (typeof value === 'function' ? value({ id, kind, name }) : value);
-  return {
-    storyId: id,
-    name,
-    component: {
-      name: kind,
-      displayName: kind.split(/\||\/|\./).slice(-1)[0],
-    },
-    parameters:
-      chromatic &&
-      CHROMATIC_PARAMETERS.reduce(
-        (acc, key) => (chromatic[key] ? { ...acc, [key]: param(chromatic[key]) } : acc),
-        {}
-      ),
-  };
-}
-
 export const extract = global => {
   const { __STORYBOOK_CLIENT_API__ } = global;
 
@@ -39,22 +12,20 @@ export const extract = global => {
 
   // Storybook 5+ API
   if (storyStore.extract) {
-    return Object.values(storyStore.extract()).map(specFromStory);
+    return Object.values(storyStore.extract());
   }
 
   // Storybook 4- API
   return __STORYBOOK_CLIENT_API__
     .getStorybook()
     .map(({ kind, stories }) =>
-      stories.map(({ name }) =>
-        specFromStory({
-          kind,
-          name,
-          parameters:
-            storyStore.getStoryAndParameters &&
-            storyStore.getStoryAndParameters(kind, name).parameters,
-        })
-      )
+      stories.map(({ name }) => ({
+        kind,
+        name,
+        parameters:
+          storyStore.getStoryAndParameters &&
+          storyStore.getStoryAndParameters(kind, name).parameters,
+      }))
     )
     .reduce((a, b) => [...a, ...b], []); // flatten
 };
